requirements.txt
python-dotenv==0.21.0
fastapi==0.85.1
pydantic==1.10.2
uvicorn==0.19.0
python-multipart==0.0.5
pyngrok==5.1.0
numpy==1.22.1
matplotlib==3.6.2
pandas==1.5.1
keras==2.10.0
opencv-python==4.6.0.66
tensorflow==2.10.0
autopep8==2.0.0